+++ src/Exceptions/UserHasNoBuiltInRolesException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Dlnsk\HierarchicalRBAC\Exceptions;
+
+class UserHasNoBuiltInRolesException extends \Exception
+{
+
+}